API completa!

+++ ecommerce/src/main/java/br/org/serratec/ecommerce/repositories/ItemPedidoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.org.serratec.ecommerce.entities.ItemPedido;
 
-public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Integer> {
+public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long> {
 
 }
